% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnetvis.R
\name{netvis}
\alias{netvis}
\title{Visualise route networks}
\usage{
netvis(
  x,
  min_width = 2,
  max_width = 6,
  width_regex = "bi|du",
  width_var_name = NULL,
  output = "leaflet",
  width_multiplier = NULL,
  ptile = 0.95,
  popup_vars = NULL,
  basemaps = NULL,
  ...
)
}
\arguments{
\item{x}{}

\item{min_width}{}

\item{max_width}{}

\item{popup_vars}{Variables to appear in popups}
}
\value{
A tmap or leaflet object
}
\description{
Visualise route networks
}
\examples{
rnet = rnet_central
netvis(rnet)
netvis(rnet, min_width = 3)
# Small variations in line width
netvis(rnet, min_width = 3, max_width = 10)
rnet = rnet_limerick
netvis(rnet, width_regex = "Bicycle")
m = netvis(rnet, width_regex = "Bicycle", output = "tmap")
class(m)
tmap::tmap_mode("view")
m
popup_vars = c(
  "Cycle friendliness" = "Quietness",
  "Gradient" = "Gradient"
)
netvis(rnet, width_regex = "Bicycle", popup_vars = popup_vars)
netvis(rnet, width_regex = "Bicycle", popup_vars = popup_vars,
   width_var_name = "Bicycle trips")
pal = c('#882255','#CC6677', '#44AA99', '#117733')
quietness_breaks = c(0, 35, 70, 85, 100)
basemaps = c(
  `Grey basemap` = "CartoDB.Positron",
  `Coloured basemap` = "Esri.WorldTopoMap",
  `Cycleways (OSM)` = "https://b.tile-cyclosm.openstreetmap.fr/cyclosm/{z}/{x}/{y}.png",
  `Satellite image` = "https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}'"
)
netvis(rnet, width_regex = "Bicycle", popup_vars = popup_vars,
   width_var_name = "Bicycle trips", col = "Quietness", pal = pal,
   breaks = quietness_breaks, basemaps = basemaps)
}
