% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnetvis.R
\name{netvis}
\alias{netvis}
\title{Visualise route networks}
\usage{
netvis(
  rnet,
  min_width = 2,
  max_width = 6,
  width_regex = "bi|du",
  output = "list",
  width_multiplier = NULL,
  ptile = 0.95
)
}
\arguments{
\item{rnet}{}

\item{min_width}{}

\item{max_width}{}

\item{attrib}{}
}
\value{
A tmap or leaflet object
}
\description{
Visualise route networks
}
\examples{
rnet = rnet_central
netvis(rnet)
netvis(rnet, min_width = 3)
# Small variations in line width
netvis(rnet, min_width = 3, max_width = 10)
rnet = rnet_limerick
netvis(rnet, width_regex = "Bicycle")
}
