% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnetvis.R
\name{netvis}
\alias{netvis}
\title{Visualise route networks}
\usage{
netvis(
  rnet,
  min_width = 1,
  max_width = 10,
  width_regex = "bi|du",
  output = "list",
  width_multiplier = NULL,
  ptile = 0.95
)
}
\arguments{
\item{rnet}{}

\item{min_width}{}

\item{max_width}{}

\item{attrib}{}
}
\value{
A tmap or leaflet object
}
\description{
Visualise route networks
}
\examples{
rnet = rnet_central
library(dplyr)
library(tmap)
netvis(rnet_central)
netvis(rnet_central, min_width = 3, max_width = 6)
}
